windows systray: fixed newlines, file locations and names
