--- win/tray/__init__.py
@@ -1,1 +1,0 @@
-# empty, allow python to walk trough
